refactor (LottoPrize) : LottoPrize를 리팩토링 한다
- lotto 패키지로 이동
- 숫자의 가독성 증가
